Use peek instead of backup()
This allows us to check for col/lines.
e.g.
if you advance the lexer to a newline, we do line++, col = 1 (first col)
but if we allow backup, we need to line--, but we already lost the
previous column number

we could, for example, save the previous column, but that would
(potentially) mean we need to save *ALL* of the column size information

So instead, we use peek() to check the next rune, and only call advance
when we need to. this means we can worry about advancing the line and
column numbers only