Change how tokens are emitted
* l.emit() is now responsible for creating a token, as well as
  consuming the input string (and copy it in the Token itself)
* l.emit() is now responsible for assigning important information
  such as the position of the token when emitting
* l.emit() also does some post-lexing cleanup, such as unescaping

Because of the above, the parser can now know about the position
of the tokens by querying the tokens themselves. This means that
when we encounter a parse error, we can easily embed this information
in the error object.

A new error, ParseError, was created For this purpose. It gives the
receiver a clear message as to the line and column number where the
error occurred.