Change how statements are deployed
macopy pointed out:
default valueに `;` 入ってたら詰む？？？
単純にセミコロンで区切るのだと

So very true. Therefore in this commit we take the parsed result
from schema and use that to separate ut each statement. This requires
that the diff package gives us a way to get list of statements, and
not a string with all statements lumped up into a single string.
To allow this, a new function diff.Diff has been introduced